Adding code for kill event for flowControl threads. Fixes #93
Currently, the flowControl threads are not properly terminated
along with dfk.cleanup and this creates a situation where the
strategy methods are called event after the DFK has shutdown resources.
This commit adds an kill event that the timer thread waits on,
which can be set by the parent at cleanup time.