Fix D400 docstring errors
D400 states that first line of the docstring should end with a period.
In this commit, I have tried to be as non-invasive as possible and
include the minimum changes necessary to fix the error. In a few cases
this required re-ordering sentences, because the first sentence should
be an imperative summary. The only other changes included are minor typo
fixes which should be non-controversial.