Tweak documentation organization
This tweaks the documentation organization slightly to add an API
reference for all user-facing classes. To provide links to the class
documentation that is not overwhelming, I used `sphinx-autogen api.rst`
to automatically create individual stubs for the user-facing classes (so
they do not cross-reference the dev docs), and only point to a single
class file so as not to be overwhelming to users. This required adding
`:noindex:` in a few places in the dev docs in order to tell Sphinx not
to cross-reference there. This commit also adds inter-documentation
cross-referencing: now parsl classes can be cross-referenced if they are
surrounded with backticks. Eventually we should update the docstrings to
take advantage of this. Fixes #115.