Add a few details to the contributing doc
This clarifies that hot fixes can be pushed to master, and that
we do not need an issue for very minor changes. Fittingly, I did
not open an issue for this. The main idea here is that issues
are useful if we need to discuss and plan BEFORE development
happens. For very minor, quick fixes, it is more work to explain
the change you want to make than just make it and discuss in the PR.