Add Parsl testing plugin
The purpose of this plugin is to cycle through configurations in
`tests/configs` and run each test on each config. This forces us to make
good on out promise that the code should be factorizable from the
configuration. Doing this turned out to be much more straightforward in
pytest, so I switched to that from nose. Because various configs are
being switched in and out by pytest, I needed to add a `clear` method
for the `DataFlowKernelLoader` to reset the DFK to `None` so that a new
DFK can be loaded without raising an error.