Remove redundancies in AALST patterns
This speeds the AALST tests up from 27s -> 7s (with an additional test
which was previously being run in `__main__` and shortens the code
from 12229 lines -> 433 lines. It also renames an app from
`test_xor_split` to `xor_split` to remove a pytest warning that it
cannot run the app directly.

I am temporarily disabling these tests for now. They take a long time
and should be modified such that they assert something.