Change checkpoint file modes to append. Fixes issue #230.
Previous behaviour truncated and overwrote the checkpoint
file at each checkpoint, although it was intended to
append.

The previous mode, wb+, means to open the file for both reading and
writing, truncating if necessary - some documentation refers to
this as "updating", which may be the source of confusion.

The new mode, ab, means to open the file for writing, but without
truncating, and positioned initially at the end of the file.