Set `GIT_WORK_TREE`
This solves a problem reported by @benclifford, where clean/dirty was
not being set correctly if the PWD is outside the parsl repo.

An additional change introduced here is to catch any exceptions when
trying to determine the code state; if one is raised it will be logged
along with the stacktrace, and the version returned will be
parsl.__version__ (with no additional information about the code state.)

A final minor change is that I removed a redundant `else` clause.