Skip check for install location
Return the same version string regardless of install method. This
ensures that if the user has modified the code and then run `python
setup.py install`, it will be reflected in the version. This also
ensures that if the printed version matches parsl.__version__, we
know there was an exception raised when determining the code state.
Caught by @yadudoc.