Add a singleton class for loading the DFK
This allows users to load a configuration which will be used to
instantiate a DFK, which is set as the active DFK. After a DFK has been
loaded, Apps can be imported without specifying an executor. If an app
does not specify an executor and a DFK has not been loaded, a
`RuntimeError` will be raised. Fixes #50.