Automate docker tests
This adds a fixture which will be run once at the beginning of tests, if
the tester has docker installed, to set up the images. It adds a basic
docker image with parsl to the configs directory such that by default
all tests will be run on docker (if the user has docker installed). The
`test_docker_multisite.py` tests have a local decorator and so will only
be run once on the multisite config. These changes make `test_docker`
and `test_without_containers` redundant, so I have removed them.

The actual test in `test_docker_multisite` is currently failing; I am
skipping it temporarily and will investigate later.