Fix `test_sites` to work with pytest
The main change here is to refactor the code bits and the site bits. The
configurations that were being tested have all been moved to the
`configs` directory. Any user-env-specific bits must be configured in
`runtime.py`; if they are not, that site will be skipped. Currently I
have some hints in there from @yadudoc's setup. Most of the tests were
redundant with tests which are already in `test_python_apps`; such tests
have been removed. For the remaining tests, I've added a `config`
command line option that can be passed in order to select which site
config to run. I have also renamed `runtime` to `user_opts` for clarity.