Remove plugin bits for running tests in subprocess
This removes a decorator which could be used to indicate a test should
be run in a subprocess. This was needed when we had tests depending on
checking the logs for various things; this sometimes broke in pytest
when running multiple tests concurrently. Now that it is not needed, I
am removing it, but not squashing it, in case we ever need to revive it.
Let us hope not: it is slow and adds complication.