Move tests that should be run by hand
I wanted to keep these separate because they are not meant to be run
with pytest, but I do not like having them out of the testing directory.
Maybe this is a good compromise: they are in the testing directory, but
ignored. Maybe at some point we can think of a better name than
`integration`. Note I had to use a somewhat clunky way of ignoring the
directory; I do not think we have access to the full path, so I had to
put a few possibilities depending on where you are calling pytest from.