Hardcode rundir
In general I have been putting rundirs in unique directories so that
multiple tests can run concurrently, but here we need to hardcode it
because we are running part of the test in a different process. Also:
disable test which fails intermittently according to machine load and
fix some broken paths.