Add more error handling in usage tracker
This commit broadens exception handling to catch exceptions from
anywhere in the UDP sender.

It also validates the values of UDP IP and port.

The `done` variable did not exist, and was generating an exception
when used. This commit removes use of it.

When an error is encountered, a compact log message is generated,
and the error is propagated no further.