Add `local` config to `basic` configurations
The special config `local` informs pytest not to try to substitute in a
new config, but rather to run with whatever the test's locally-defined
one is. This is needed for certain tests which are testing, for example,
the `cleanup` behavior, or other behavior where the test itself needs to
control the dfk without interference from pytest. The aim of the `basic`
command line option is to provide a reasonable choice in the trade-off
between test running time and comprehensiveness of the tests run. I
think `local` tests should be run to ensure adequate comprehensiveness.