Add a RepresentationMixin class
This class is needed in order to provide a representation (for logging,
etc) of class-based configurations. This added complexity is an
unfortunate side effect of moving away from json-based configs. On the
other hand, I think this class will be useful in general for
providing representations of classes which previously had to define
their own __repr__ method.