Changes to Azure provider for class-based configs
This provider is not functional. I will open a separate issue about
that. For now I have just made superficial changes in the spirit of
class-based configs. I have also fixed some obvious bugs where I found
them, but do not want to get caught in the weeds of testing this
provider.