Split up travis tests
I have removed a few stale testing files and switched to running each
directory or root-level testing module separately with Travis. This is less
convenient, but seems to be slightly less flaky. We should continue
to improve test robustness and until we can go back to the usual pytest
invocation.