Make sure that `dep_fail` tasks are retried
This required making two static methods into normal methods. I didn't
see the static functionality being used anywhere, so I think this should
be ok. Instead of relying on whether a future has had an exception added
to determine the state, it is now read it off of the state table directly.
Otherwise, futures that are being retried will be counted as finished.
Fixes #473.