Fix a couple of places exception handling was awkward.
In the first case, an exception from an application is not
forwarded to the submit side, because of an incorrect logging
statement. This patch fixes that logging.

In the second case, an exception (actually the above logging
statement, when I discovered this) has its description printed,
but no stack trace information. This patch re-raises the exception
so python logs the stack trace.