Put all run-related files in the run_dir
Unless otherwise configured, this makes all run-related files go into
the run_dir (fixes #457). It also eliminates the script_dir keyword
argument in the Provider (instead, it is hard-coded to be inside the
run_dir); fixes #442. Having it configurable adds no benefit and a lot
of confusion. By default, if the local run_dir is accessible on the
remote machine, output files will be put there; otherwise they will be
put in a run_dir with the same number on the remote side. This should
also fix Parsl/libsubmit#74 (correspondingly, fix #521).