Count future readiness by asking the futures
Prior to this commit, dependencies were counted by looking at task
status.

After this commit, dependencies are counted by asking the futures that
represent the dependencies.

I think that this does not change behaviour of the code now, but makes
the code behave more consistently:

Some code assumes that when _count_deps has returned 0, all futures are
ready and we be sure that potentially blocking actions on dependency
futures will not block.

An example of this is in sanitize_and_wrap.

This commit moves towards doing all dependency handling via the
relevant futures, rather than inspecting the task array.

Testing: as of 2018-10-25, all tests pass on my laptop