Remove spurious exception in handle_update non-lazy-error path.
When handling an exception in non-lazy_errors mode, handle_update
would throw that exception itself. The lazy_errors mode would
instead return normally.

In some circumstances, when a task ran extremely fast, an exception
thrown in an app would (because of the above) turn into an exception
thrown in handle_update, which was then unexpectedly thrown to
(I think) the add_done_callback site (as the future was already
complete at the time of callback).

This commit removes that specific exception, but does not
protect the guts of parsl against other exceptions thrown in a
callback handler.

This addresses something to do with issue #312 (but maybe not
exactly issue #312)