Catch exceptions when adding done callbacks
In some circumstances, an exception thrown by a
future done callback is passed out of the call
registering that callback - sometimes when the
callback happens immediately, so is called
within the call stack of the registering call.

This commit attempts to catch all such 
exceptions, log them for interest, and then
ignore them - as should happen if the callback
is made asynchronously.