Remove implicit wait for outstanding tasks at DFK shutdown.
In practice, this turned out to be awkward when, for example, ctrl-C
was pressed and parsl spent a long time waiting (or sometimes hung).

This commit intends that the same wait code is executed explicitly
at the place in the workflow code that the wait is desired (usually
the end). Ctrl-C and other abnormal exits will not cause this wait
to happen.