Attempt to enforce the presence of a label on ExecutionProviders
Previously, there was no requirement that an ExecutionProvider have
a label, yet some code assumes there always is a label.

This modifies all providers which are ExecutionProviders to have
such a property (although I have not tested them all).

Some providers in this directory are not ExecutionProviders, and I
have not touched those: googlecloud, jetstream.

This commit also introduces a potential bugfix when using the kubernetes
provider, which did not implement a label property.

This commit makes mypy happy that labels are used properly.