Remove parent member from AppFuture implementation
This put all knowledge of the parent(s) of an AppFuture
into a single place - the DFK - rather than having it
split between the DFK and the AppFuture.

The AppFuture parent_callback is now invoked by the
DFK in the callback that the DFK has registered with
parent futures, rather than the parent futures having
multiple registrations. This makes the result handling
codepath driven through that one single callback.

Code relating to the parent attribute of AppFuture
is removed. This includes removing the overridden
running() method, which was not implemented in any
existing parent future implementations; and removing
app future parent references from DataFuture.

This will possibly change the order in which callbacks
happen - and subtle races there might be exposed. This
would most likely be exposed to an end-user if they have
registered callbacks with an AppFuture which expect the
DFK task state to be in a particular state; the order
of state update vs AppFuture callback invocation may
now be different.