Make memo-based callback look like execution-based callback
This makes the memo-based callback happen via the (memo)
Future, in the same way that the execution-based callback
happens via an execution Future.

This weakens the requirement that memo_fu have
a result already at the point that this piece of
code is reached: if it doesn't, the callback will
be deferred until it does.

In practice, this should not change user-exposed memo
behavior - instead this is to simplify code.

This is a step towards making memo- and execution-based
completion callbacks be handled in entirely the same
way, as part of pulling out pieces of PR #722.