Combine two exec callbacks into one
Previously the exec_done callback was added in two
places in launch_tasks, depending on whether the
execution path or the memoization path was taken.

This commit moves that callback addition to outside
of launch_tasks, where we do not need to care where
the launched task future has come from.