Add script_path parameter to Launchers, especially SrunLauncher
This means that launchers can create their temporary files in the run
directory, rather than a more arbitrary directory such as pwd.

SrunLauncher is the only launcher that this commit actually changes
to use the script_path, as that was the one used in the motivating
imsim work.

TODO: This raises an issue (possibly also encountered by @gordonwatts
with the ssh provider?) of which file systems are being used for
creating launcher files - the local or the remote.

This is based on work done for imsim.