Support for identifying managers by block_id, and putting managers in a block on hold
Managers are given a block identifier at initialization that is reported back at registration.
This allows the executor to track tasks/worker slots at the block level.
The executor also has a _hold_block() function that puts all managers in a block on hold
preventing further task scheduling, and making it safe to terminate the block.