Increase the number of tasks in htex CI test
There was previously a bug where this test would fail after a
relatively small number of tasks executed, and so this test
was introduced to CI with a small number of tasks.

This bug was fixed in PR #725, and so a more intensive test can be
run now - this commit makes that change.