Fix flake8 error introduced with flake8 3.7
This fixes 8 instances of:

F632 use ==/!= to compare str, bytes, and int literals

This will introduce a potential behaviour change: previously, `is`
was used to compare strings; now `==` is used to compare strings.

Where the strings have been interned (which I hope has been always),
then the behaviour should be the same; but there are potential cases
(at least abstractly) where `is` was returning false for strings that
were ==-equal, and `is not` was returning true for string that were
==-equal. So if you've bisected a problem back to this commit,
check that out!