Add clean network marker in tests
At least one test that fails in CI works on a host with a clean
network connection - the Travis CI network is behind a NAT and does
not work properly with, for example, FTP.

That test is marked with a new cleannet mark. It will not run in CI,
but can be run from manually chosen hosts (or a different CI one day
perhaps)

The help test for markers is updated with brief info.

The documentation on how to get marker help text is updated to
give the parsl-specific help text. Previously it did not include
parsl-specific information.