Add a command line for explicitly authorising to globus
... because otherwise that doesn't happen until something
happens, and stdin/stdout/the user might not be available
then.

For example, in the test suite, no stdin is available;
in a big workflow running on a cluster worker node, there
won't be interactive access to that workflow, which will
likely run at some arbitrary point long after submission.