Remove some n^2 scaling in interchange (#728)
Prior to this PR, the interchange inspects every worker registration for every task message received.

This means when allocating n tasks to n workers, this is O(n^2)

In practice with imsim in production this was producing noticeable slowdowns.

This PR changes the behaviour so that a set of interesting workers is maintained, and only interesting workers are inspected when allocating a task. Interesting workers are the workers which have available capacity.

In the imsim work which motivated this PR, there are on the order of 2000 workers, and a larger number of long running (many hour) jobs. Testing and thinking-about should happen for different workloads (for exmaple, the case where workers are starved of work, and when jobs are very short.