Make File not a subclass of `str` - a breaking user-facing API change
File was originally made a subclass of `str` to attempt to allow File
objects to be used where file paths are expected - for example, with
open().

Python 3.6 introduces a proper way to do this, using the __fspath__
protocol; and File already implements this.

Python 3.5 does not support __fspath__, and only `str` or `byte`
objects can be passed to `open`.

Prior to this commit, File attempted to override the implementation
of __str__ to dynamically provide a posix path.

However, this does not always work: `open` uses the immutable superclass
string value, not the value returned by `__str__`.

When a `File` is constructed like this: `File("./out.txt")` then the
immutable str value is (probably) correct and no error occurs. The
existing test cases catch this.

When a `File` is constructed like this: `File("file:///tmp/out.txt")`
then the immuatable str value is that entire URI, even though invoking
__str__() returns a reasonable posix path: `/tmp/out.txt`

This case of opening a URI-based File is not tested in the current test
suite.

Because the immutable str value, and the result of __str__()
do not agree, some awkward error messages arise. For example,
an error message like this is reported:

```
Exception: [Errno 2] No such file or directory: /tmp/qwertz.out
```

even though /tmp/qwertz.out is somewhere writeable, because
(according to `strace`) the OS open call was invoked with the
immutable str value:


```
open("file:///tmp/qwertz.out", O_RDWR|O_CREAT|O_APPEND|O_CLOEXEC, 0666 <un
finished ...>
```

This commit does not change behaviour for users of python >=3.6, but
requires python 3.5 users to explicitly convert a File object to
a path, using `str()`.