Make DataFutures complete after stageout, not before
WIP

It looks like data futures are marked as complete
as soon as the generating app task / AppFuture is
complete.

This could be before any corresponding stageout
task has been executed, and so be before the
file can actually be observed at the appropriate
location.

The datafuture should be looking at the stageout
job for success, not the app task.

At present the stageout task doesn't come
into existence until after the job has completed
(I think) - because stageout calls are made as
part of the app completion.

Do those stageout tasks have an AppFuture too?

Having the DFK driving DataFuture completion
might be a better place for this completion
logic? (as I was pushing at with AppFuture
refactoring)