rearrange previously disabled test that needs a clean network (#754)
* Re-enable ftp: test that was marked as broken but appears to work now

* try a different url for ftp test as IANA doesn't work from travis

* Revert "try a different url for ftp test as IANA doesn't work from travis"

This reverts commit b8888329c3d7e5e955a21977c6903247ee4572ee.

* Make a test mark for tests that require a clean network connection, and make CI build skip those tests as travis network is not clean - it has FTP-resistant NAT

* Add clean network marker in tests

At least one test that fails in CI works on a host with a clean
network connection - the Travis CI network is behind a NAT and does
not work properly with, for example, FTP.

That test is marked with a new cleannet mark. It will not run in CI,
but can be run from manually chosen hosts (or a different CI one day
perhaps)

The help test for markers is updated with brief info.

The documentation on how to get marker help text is updated to
give the parsl-specific help text. Previously it did not include
parsl-specific information.

* Mark another disabled test as enabled when the network is clean