Move module level definitions into Globus class.
Prior to this commit, importing this module caused
a write to be made to a user's home directory. In
some cases (see bug #780), a user's home directory
is not always writeable, and in those cases this
class could not be imported at all.

After this commit, such home directory write problems
should appear when attempting to use Globus transfers,
but not otherwise.