Remove indirection of staging via File objects
Prior to this, file objects pretty much just forwarded on
staging requests to the data manager without doing anything
interesting.

This removes that indirection.

It removes some of the need for a file to be associated
with a data manager - something which sits awkwardly with
wanting to reuse Files between DFKs, for example in
interactive notebook use.

Instead, each DFK has a reference to a data manager directly.