There is probably no need to make staging requests go via a file object?
This simplifies the call graph, i hope.

It removes some of the need for a file to be associated with
a data manager - something which sits awkwardly with
wanting to reuse Files between DFKs, for example in
interactive notebook use.

The remaining need for knowing a data manager is to do
with using a global Globus provider for execution-side
path resolution. This is something that a future commit
should address.