Remove duplicate setting of local_path for Globus staging
Now local_path is set in the same way as for other staging
methods, and some globus-specific special casing elsewhere
in the code is removed.

The way in which self.globus is initialised has to move around
because of this change, and now happens in an explicitly
named method.

The globus-specific list of files in the data manager, which
exists only for an undone TODO, is removed.