Potential bug? I'm not sure and should try to make test cases.
File objects are shared between executors; but file objects contain
a field for the remote executor path. That remote executor path
will be different for different executors. But file objects cannot
express that.

I think this should be testable, and this patch is for me to work on
testing it. For example, using two threadlocal executors would let it
run on one machine; perhaps with http staging in of the same file to
two different executors?

Something that is relevant here is *where* the staging happens:
sometimes the File object will be shared (eg with ThreadLocal)
between executors, and sometimes it won't. So the time of modification
is relevant... which is a bit icky... and more of an arugment to
move away from this particular representation of remote/local files.