Raise an exception rather than use wrong cwd for globus paths (#788)
Prior to this patch, if no execute-side working directory is known for a
globus transfer, the client side current directory is used.

This is very likely to be wrong (unless the execute side and client side
are running on the same filesystem, perhaps), so this patch causes an
exception to be raised asking for more explicit configuration rather
than guessing.