Allow alternate modes to be specified for bash app stdout and stderr.
Modes can be specified via a (file,mode) tuple as an alternative
option to a simple filename string.

Default mode remains "a+" for now.  Typical alternative fir simple
file overrwite is ('filename','w')

Also fixes error in exception handling for BadStfFileStream and
MissingOutputs exception.  Problem was that the exception superclass
needs to be initialized withthe same member variables as the
subclassed execption in order for it to pickle correctly in the
"remote" execution context.

Other exceptions have similar issues, but were not corrected, pending
review of the proposed fix.

The app_stdout.py pytest was extended to test the new options, and
corrected to initialize correctly (by not attempting to clear and
reset the dfk, which seemed to causing this test to file.

Note that only the local execution tests were tested at the time of
this commit.