Add a poll period parameter to htex (#741)
* attempt to parameterise poll period changed in source in imsim

this doesn't seem to work - instead there's a hard loop

maybe there's a hidden exception? or something. it's not been
obvious why it isn't working...

* set default period in test to the previously hard-coded value

* wip

* pass poll parameter to interchange, rather than accidentally to workers

* tidyup

* set test poll period back to 1ms

* Adding poll_period option to the process_worker_pool and some serious cpu utilization fixes

The default timeout is set to 10ms which surprisingly is more performant than 1ms
Push_results thread now does a blocking wait on the pending_task_queue.

* Setting poll period default to 10ms for consistency

* Draw queue size limit from manager

* Fixing issues caught by Zhuozhao