Make DFK initialisation not module-global
The test module is imported on the remote side by canning code,
as part of uncanning some lambda expression (I'm not sure
exactly which one).

This was causing a DFK to be loaded on the remote side, because
the DFK was being loaded as a module-level global.

Initialising two DFKs close in time seems to result in a race
condition creating working directories.