Set the Future-result of a DataFuture to the correct result
Prior to this, the superclass Future implementation was given
a result that was inconsistent with the values returned by
(for example) DataFuture custom implementations of result().

This leads to fragility and potential inconsistency about what
a DataFuture's result value actually is.

This change makes that value always be the wrapped File object,
which is the result that the custom implementations attempt to
return.