Use enforce.py for runtime checking of top level Config (#862)
This uses PEP484 type hints, the same annotations which mypy will statically check, but on suitably annotated functions, those type hints are enforced at runtime.

Config based type errors are the main place where this could help users - see issue #855 and PR #832

Error messages resulting from this will look like this:

E   enforce.exceptions.RuntimeTypeError: 
E     The following runtime type errors were encountered:
E          Argument 'strategy' was not of type <class 'str'>. Actual type was int.

To use this fully, all of the functions for config-like code (i.e. constructors and helper functions that you'd expect to find in a config definition) should be type-hinted and the @runtime_validation decorator added - I have opened issue #873 to track that.

Some suitable type hints might (or might not) be found in the benc-mypy branch (PR #863), where I've been shadowing the master code base but with type hints suitable for mypy.

This complements mypy usage: mypy can verify that pieces of the parsl codebase correctly call other pieces of the parsl codebase, statically; enforce can verify that other people's code calls pieces of the parsl codebase correctly, dynamically.

This PR also breaks mypy typechecking for Config - mypy does not have type information for the runtime_validation decorator, and so this PR tells mypy to ignore that decorator, meaning that decorated functions lose their type info as far as mypy is concerned. This could be addressed by providing a type stub to mypy for runtime_validation. This is a breakage of the spirit of PR #832 which was intended to allow static checking of configurations, for example ahead of submitting to a queue.