Fix broken #856 changes to interchange->executor exceptions (#860)
Revert two changes from #856: in result_package, when tid == -1, the type of exception should be Exception after all (as was the case before #856). It is only when tid != -1 that the type of exception should be RemoteExceptionWrapper.

In the third change in #856, it is necessary for an exception to be actually raised, rather than the exception object just be constructed, in order to pass it correctly with RemoteExceptionWrapper; and the parameters passed to RemoteExceptionWrapper were incorrect. This PR addresses both of those issues.

* fix exception sending - when tid == -1, 'exception' should contain a serialised exception, not a remote exception wrapper; but when tid != -1, 'exception' should contain a RemoteExceptionWrapper

* remove debugging

* more fiddling with exceptions

* remove debug changes to test

* remove error message change, to go in other PR

* fix flake8