Make scale_in parameter name consistent in all executor classes
This is needed in the now-less-obscure case of calling scale_in with
named parameters - the base class previously said that scale_in could be
called with (count=1) as a parameter but no subclasses accepted that.

The terminology now pretty much is settled as 'blocks', so that is
what this PR changes everything to.