Debug and fix local provider state machine
When a job was finished, it was not properly going into a failed
or completed state - it would only do so if the job was not in a RUNNING
state

Add in some debugging logs to help with this, and an error.

questions:
should the error message actually be a fail/throw exception to fail
harder?

how much of this logging shoudl be kept? i feel like there
should be a bit here but not this much?
or maybe this much is fine .. its interesting to see the jobs change state