Tidy debug messages around heartbeat/task requests
Issue #898 shows a confusing log message, where a heartbeat
incorrectly logs that HEARTBEAT_CODE tasks have been requested,
in addition to logging a heartbeat.

This commit moves that log into the non-heartbeat path only, so
that exactly one of the heartbeat or task request messages will be
logged.

The phrasing of the heartbeat log message is modified to be in the
same sense as the task request message.