Introduce mypy stub for typeguard, and type-check parsl.load() (#874)
This allows mypy to understand the type signatures of functions
decorated with @typeguard.typechecked, allowing the type signatures
on such functions to be used by both enforce and mypy.

Prior to this commit, adding @typechecked removed the ability of mypy to
type-check calls to decorated functions, as it did not understand that
the type of such a decorated function is the same as the original
function. The added stub makes this declaration.

The test-suite mypy calls are modified to use that stub file.

The immediate motivation for this is to keep mypy passing when adding a
type annotation for parsl.load; that type annotation is also added in
this commit.

This commit fixes some of problem described in the commit message for
d7d9a255 about breaking mypy type checking for end user Configs in some
places, but by no means all of it.

However there is a mypy bug, https://github.com/python/mypy/issues/5398,
where decorated __init__ methods are not properly type-checked by mypy,
which means that mypy typechecking is still missing on any decorated
__init__ methods - which unfortunately is most of the user-facing type
checked code.

At time of writing, there are PRs open to fix this in mypy, so there
is hope; and that does not impede this PR being merged.