Print path to stdout and stderr for failed tasks (#923)
This prints path to stdout and stderr for failed tasks. Note that this might not be as helpful in situations where there is no shared filesystem. But we have various issues describing that and related problems, and I think it's ok to keep it factorized out. In any case, this is consistent with what is saved in the monitoring DB.
Fixes #849.