implement hook points for staging that can be configured at runtime
instead of dispatching inside the DFK explicitly based on a
hard-coded map between URI schemes and staging implementations,
go through each staging scheme in sequence, giving it the chance
to handle staging - perhaps with a 'can_handle()' check which
needs to behave consistently across multiple calls because it
will get called at each point?

Then http and ftp need to be put into their own (default)
staging handlers.

At this stage, the only hook that can happen is the same
place as before - stage in before job, and stage out after
job. Future commits should add in different hook points.

The concept of 'is_remote' being a property of files defined
by the URL scheme is also superceded by this, and so
is_remote is removed. For behaviour of "file:" URLs to mean
'no staging', then a no-op staging scheme should be implemented
in this PR; but it is concievable that that might be
overridden by something that knows more about the
setup of things (which would allow changing the semantics of file:
URL from "don't stage" to "I know how to stage from the
local file system" for that particular staging scheme - something
that isn't supported well at the moment - for example the
workqueue provider does it ad-hoc)

There's a user-facing API change here: if you specify staging
mechanisms as the empty list [] rather than as None or omitting
it, then the default staging mechanisms will be removed and
even no-op file access won't work.