Add __init__.py to parsl.tests.test_staging
Prior to this commit, the tests in:

* test_docs_1.py
* test_implicit_staging_ftp.py

would not successfully run in pytest, as their module names were not being
picked up by pytest correctly.

These tests were not running, and so not failing in CI, because those
tests are explictly disabled in CI using the 'cleannet' mark. Travis
does not provide a suitably clean network environment for those tests to
run.

Those tests can be run (and were failing, but now pass) on my laptop on
a residential DSL connection.