Change visualization --external to --listen
This allows arbitrary IP addresses to be bound for listening, rather
than only allowing a choice between 0.0.0.0 and 127.0.0.1

This fits more in line with how most server software allows listening
address to be chosen (see apache httpd, BIND, postgres).

As a side effect, that also allows the visualization code to listen on
IPv6.

This should fix issue #1023