Propagate the checkpointing hashsum to apps
This allows apps to persist partial state between workflow runs
in a way that is safe with checkpointing.

The app can name intermediate state based on the hashsum, and that
state will change name in the same circumstances that a checkpoint
would be invalidated.

This allows partial results living on the filesystem (such as
app-specific checkpoints) to be re-used alongside the parsl
checkpointing mechanism, rather than requiring apps to know how
to correctly distinguish in-progress state themselves.

Usage: add a checkpoint_hash=parsl.AUTO_LOGNAME parameter, and
turn on caching.

    @App('bash', cache=True)
    def foo(x, checkpoint_hash=parsl.AUTO_LOGNAME, stdout=None):

That checkpoint_hash parameter will be replaced at invocation with
the relevant checkpoint hash key.