Channels are optional. Bodge this.
A better approach might be to
have a Channelable protocol and check
that we have the right instances of that?
(with isinstance())

Or enforce a Channel attribute which is
sometimes None. But that's ugly because
then even things which know they have
a channel for sure (such as slurm)
need to deal with the possibility that the
channel has later on been set to None.