Add per-monitoring-worker logging
Each worker will log to a file in /tmp, the same places that
bash per-task log files are sent. My intention is that this will
become customisable in a later PR.

Debug level messages are placed around steps that have the potential
to take a long time.

The log level is set to DEBUG, so potentially large amounts
of log information may be generated (several lines per monitoring
loop iteration per worker). This is because the monitoring code
generally needs debugging at the moment. This should also
become parameterisable in a future PR.