Flush stderr after printing executable diagnostic
In some cases, this message wasn't being sent to stderr
before the stderr of the spawned executable, meaning the
executable diagnostic information appeared at the end,
rather than start, of the logged stderr.

This added flush should provide the necessary sequencing
to ensure the message appears before the stderr of the
spawned process.

See issue #1008