Allow monitoring db task_time_submitted to be NULL
When a task is written to the DB which has not been submitted, then
prior to this commit, a database integrity error was raised, as the
database required a task to have a submit time.

This manifests in #1100 where unsubmitted tasks are written out to
the monitoring DB at the point that a workflow fails.