Remove flake8 exemption for F403 undefined name detection
When a source file contains an `import *`, flake8 cannot
check that source file for use of symbols which are
not defined.

This commit also fixes some problems that were covered
over by the F403 exemption.

A couple of files have other problems which can have a
tighter per-file exemption applied.